oops, incomplete long policy
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@5841 a1c6a512-1295-4272-9138-f99709370657

+++ apps/plugin.h
@@ -191,7 +191,7 @@ struct plugin_api {
     void (*sleep)(int ticks);
     void (*yield)(void);
     long* current_tick;
-    int (*default_event_handler)(int event);
+    int (*default_event_handler)(long event);
     int (*default_event_handler_ex)(long event, void (*callback)(void *), void *parameter);
     int (*create_thread)(void* function, void* stack, int stack_size, const char *name);
     void (*remove_thread)(int threadnum);
